refactor : ui controller와 api controller 구분 #39
