+++ src/main/java/jpabook/jpashop/controller/api/common/response/UserAccountResponse.java
@@ -1,4 +1,4 @@
-package jpabook.jpashop.controller.common.response;
+package jpabook.jpashop.controller.api.common.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jpabook.jpashop.enums.user.account.CashFlowStatus;
