+++ src/main/java/jpabook/jpashop/controller/user/UserAccountController.java
@@ -5,9 +5,9 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jpabook.jpashop.aop.annotations.Loggable;
-import jpabook.jpashop.controller.common.annotations.LoginedUserUid;
-import jpabook.jpashop.controller.common.request.UserAccountRequest;
-import jpabook.jpashop.controller.common.response.UserAccountResponse;
+import jpabook.jpashop.controller.api.common.annotations.LoginedUserUid;
+import jpabook.jpashop.controller.api.common.request.UserAccountRequest;
+import jpabook.jpashop.controller.api.common.response.UserAccountResponse;
 import jpabook.jpashop.dto.AccountDto;
 import jpabook.jpashop.exception.common.CannotFindEntityException;
 import jpabook.jpashop.exception.user.account.InvalidBalanceValueException;
