+++ src/main/java/jpabook/jpashop/filter/CustomAuthenticationFilter.java
@@ -5,9 +5,9 @@ import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import jpabook.jpashop.controller.common.request.UserRequest;
-import jpabook.jpashop.controller.common.response.ErrorResponse;
-import jpabook.jpashop.controller.common.response.UserResponse;
+import jpabook.jpashop.controller.api.common.request.UserRequest;
+import jpabook.jpashop.controller.api.common.response.ErrorResponse;
+import jpabook.jpashop.controller.api.common.response.UserResponse;
 import jpabook.jpashop.dto.UserDto;
 import jpabook.jpashop.exception.user.UserExceptonMessages;
 import jpabook.jpashop.util.JwtTokenProvider;
