+++ src/main/java/jpabook/jpashop/security/CustomAuthenticationEntryPoint.java
@@ -1,10 +1,9 @@
 package jpabook.jpashop.security;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import jpabook.jpashop.controller.common.response.ErrorResponse;
+import jpabook.jpashop.controller.api.common.response.ErrorResponse;
 import jpabook.jpashop.exception.user.UserExceptonMessages;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.AuthenticationException;
