+++ src/test/java/jpabook/jpashop/controller/user/UserAccountControllerTest.java
@@ -3,13 +3,12 @@ package jpabook.jpashop.controller.user;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import jpabook.jpashop.controller.common.request.UserAccountRequest;
-import jpabook.jpashop.controller.common.response.UserAccountResponse;
+import jpabook.jpashop.controller.api.common.request.UserAccountRequest;
+import jpabook.jpashop.controller.api.common.response.UserAccountResponse;
 import jpabook.jpashop.domain.user.Account;
 import jpabook.jpashop.enums.user.account.CashFlowStatus;
 import jpabook.jpashop.enums.user.account.CashFlowType;
 import jpabook.jpashop.repository.AccountRepository;
-import jpabook.jpashop.testUtils.TestDataUtils;
 import jpabook.jpashop.util.JwtTokenProvider;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
